backend/requirements.txt
fastapi
httpx
uvicorn[standard]
pydantic[email]
python-jose[cryptography]
sqlalchemy>=2.0
psycopg2-binary
